Added include of sys/param.h to ensure that the BSD variable exists (to properly define the parameter for scandir ).
CVS patchset: 1547
CVS date: 2003/01/07 20:43:06

+++ linux-ha/portability.h
@@ -38,6 +38,7 @@
 #endif
 
 
+#include <sys/param.h>
 #ifdef BSD
 #	define SCANSEL_CAST	(void *)
 #else
@@ -139,6 +140,6 @@ inet_pton(int af, const char *src, void *dst);
 #endif
 
 #ifndef HAVE_NFDS_T 
-	typedef unsigned long int nfds_t;
+	typedef unsigned int nfds_t;
 #endif
 #endif /* !PORTABILITY_H */
